+++ crates/agent/src/tools/edit_file_tool.rs
@@ -276,14 +276,9 @@ impl AgentTool for EditFileTool {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::{
-        ContextServerRegistry, Templates, ToolInputSender, WriteFileTool, WriteFileToolInput,
-    };
-    use acp_thread::Diff;
+    use crate::{ContextServerRegistry, Templates, ToolInputSender};
     use fs::Fs as _;
-    use futures::StreamExt as _;
     use gpui::{AppContext as _, TestAppContext, UpdateGlobal};
-    use language::language_settings::FormatOnSave;
     use language_model::fake_provider::FakeLanguageModel;
     use project::ProjectPath;
     use prompt_store::ProjectContext;
@@ -294,59 +289,8 @@ mod tests {
     use util::rel_path::{RelPath, rel_path};
 
     #[gpui::test]
-    async fn test_streaming_edit_create_file(cx: &mut TestAppContext) {
-        let (_edit_tool, write_tool, _project, _action_log, _fs, _thread) =
-            setup_test(cx, json!({"dir": {}})).await;
-        let result = cx
-            .update(|cx| {
-                write_tool.clone().run(
-                    ToolInput::resolved(WriteFileToolInput {
-                        path: "root/dir/new_file.txt".into(),
-                        content: "Hello, World!".into(),
-                    }),
-                    ToolCallEventStream::test().0,
-                    cx,
-                )
-            })
-            .await;
-
-        let EditFileToolOutput::Success { new_text, diff, .. } = result.unwrap() else {
-            panic!("expected success");
-        };
-        assert_eq!(new_text, "Hello, World!");
-        assert!(!diff.is_empty());
-    }
-
-    #[gpui::test]
-    async fn test_streaming_edit_overwrite_file(cx: &mut TestAppContext) {
-        let (_edit_tool, write_tool, _project, _action_log, _fs, _thread) =
-            setup_test(cx, json!({"file.txt": "old content"})).await;
-        let result = cx
-            .update(|cx| {
-                write_tool.clone().run(
-                    ToolInput::resolved(WriteFileToolInput {
-                        path: "root/file.txt".into(),
-                        content: "new content".into(),
-                    }),
-                    ToolCallEventStream::test().0,
-                    cx,
-                )
-            })
-            .await;
-
-        let EditFileToolOutput::Success {
-            new_text, old_text, ..
-        } = result.unwrap()
-        else {
-            panic!("expected success");
-        };
-        assert_eq!(new_text, "new content");
-        assert_eq!(*old_text, "old content");
-    }
-
-    #[gpui::test]
     async fn test_streaming_edit_granular_edits(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "line 1\nline 2\nline 3\n"})).await;
         let result = cx
             .update(|cx| {
@@ -372,7 +316,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_edit_multiple_edits(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) = setup_test(
+        let (edit_tool, _project, _action_log, _fs, _thread) = setup_test(
             cx,
             json!({"file.txt": "line 1\nline 2\nline 3\nline 4\nline 5\n"}),
         )
@@ -410,7 +354,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_edit_adjacent_edits(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) = setup_test(
+        let (edit_tool, _project, _action_log, _fs, _thread) = setup_test(
             cx,
             json!({"file.txt": "line 1\nline 2\nline 3\nline 4\nline 5\n"}),
         )
@@ -448,7 +392,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_edit_ascending_order_edits(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) = setup_test(
+        let (edit_tool, _project, _action_log, _fs, _thread) = setup_test(
             cx,
             json!({"file.txt": "line 1\nline 2\nline 3\nline 4\nline 5\n"}),
         )
@@ -486,8 +430,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_edit_nonexistent_file(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
-            setup_test(cx, json!({})).await;
+        let (edit_tool, _project, _action_log, _fs, _thread) = setup_test(cx, json!({})).await;
         let result = cx
             .update(|cx| {
                 edit_tool.clone().run(
@@ -519,7 +462,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_edit_failed_match(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "hello world"})).await;
         let result = cx
             .update(|cx| {
@@ -548,7 +491,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_early_buffer_open(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "line 1\nline 2\nline 3\n"})).await;
         let (mut sender, input) = ToolInput::<EditFileToolInput>::test();
         let (event_stream, _receiver) = ToolCallEventStream::test();
@@ -585,43 +528,8 @@ mod tests {
     }
 
     #[gpui::test]
-    async fn test_streaming_path_completeness_heuristic(cx: &mut TestAppContext) {
-        let (_edit_tool, write_tool, _project, _action_log, _fs, _thread) =
-            setup_test(cx, json!({"file.txt": "hello world"})).await;
-        let (mut sender, input) = ToolInput::<WriteFileToolInput>::test();
-        let (event_stream, _receiver) = ToolCallEventStream::test();
-        let task = cx.update(|cx| write_tool.clone().run(input, event_stream, cx));
-
-        // Send partial with path but NO mode — path should NOT be treated as complete
-        sender.send_partial(json!({
-            "path": "root/file"
-        }));
-        cx.run_until_parked();
-
-        // Now the path grows and mode appears
-        sender.send_partial(json!({
-            "path": "root/file.txt",
-            "mode": "write"
-        }));
-        cx.run_until_parked();
-
-        // Send final
-        sender.send_full(json!({
-            "path": "root/file.txt",
-            "mode": "write",
-            "content": "new content"
-        }));
-
-        let result = task.await;
-        let EditFileToolOutput::Success { new_text, .. } = result.unwrap() else {
-            panic!("expected success");
-        };
-        assert_eq!(new_text, "new content");
-    }
-
-    #[gpui::test]
     async fn test_streaming_cancellation_during_partials(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "hello world"})).await;
         let (mut sender, input) = ToolInput::<EditFileToolInput>::test();
         let (event_stream, _receiver, mut cancellation_tx) =
@@ -652,7 +560,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_edit_with_multiple_partials(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) = setup_test(
+        let (edit_tool, _project, _action_log, _fs, _thread) = setup_test(
             cx,
             json!({"file.txt": "line 1\nline 2\nline 3\nline 4\nline 5\n"}),
         )
@@ -714,47 +622,8 @@ mod tests {
     }
 
     #[gpui::test]
-    async fn test_streaming_create_file_with_partials(cx: &mut TestAppContext) {
-        let (_edit_tool, write_tool, _project, _action_log, _fs, _thread) =
-            setup_test(cx, json!({"dir": {}})).await;
-        let (mut sender, input) = ToolInput::<WriteFileToolInput>::test();
-        let (event_stream, _receiver) = ToolCallEventStream::test();
-        let task = cx.update(|cx| write_tool.clone().run(input, event_stream, cx));
-
-        // Stream partials for create mode
-        sender.send_partial(json!({}));
-        cx.run_until_parked();
-
-        sender.send_partial(json!({
-            "path": "root/dir/new_file.txt",
-            "mode": "write"
-        }));
-        cx.run_until_parked();
-
-        sender.send_partial(json!({
-            "path": "root/dir/new_file.txt",
-            "mode": "write",
-            "content": "Hello, "
-        }));
-        cx.run_until_parked();
-
-        // Final with full content
-        sender.send_full(json!({
-            "path": "root/dir/new_file.txt",
-            "mode": "write",
-            "content": "Hello, World!"
-        }));
-
-        let result = task.await;
-        let EditFileToolOutput::Success { new_text, .. } = result.unwrap() else {
-            panic!("expected success");
-        };
-        assert_eq!(new_text, "Hello, World!");
-    }
-
-    #[gpui::test]
     async fn test_streaming_no_partials_direct_final(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "line 1\nline 2\nline 3\n"})).await;
         let (mut sender, input) = ToolInput::<EditFileToolInput>::test();
         let (event_stream, _receiver) = ToolCallEventStream::test();
@@ -776,7 +645,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_incremental_edit_application(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, project, _action_log, _fs, _thread) = setup_test(
+        let (edit_tool, project, _action_log, _fs, _thread) = setup_test(
             cx,
             json!({"file.txt": "line 1\nline 2\nline 3\nline 4\nline 5\n"}),
         )
@@ -873,7 +742,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_incremental_three_edits(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, project, _action_log, _fs, _thread) =
+        let (edit_tool, project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "aaa\nbbb\nccc\nddd\neee\n"})).await;
         let (mut sender, input) = ToolInput::<EditFileToolInput>::test();
         let (event_stream, _receiver) = ToolCallEventStream::test();
@@ -958,7 +827,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_edit_failure_mid_stream(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, project, _action_log, _fs, _thread) =
+        let (edit_tool, project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "line 1\nline 2\nline 3\n"})).await;
         let (mut sender, input) = ToolInput::<EditFileToolInput>::test();
         let (event_stream, _receiver) = ToolCallEventStream::test();
@@ -1048,7 +917,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_single_edit_no_incremental(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, project, _action_log, _fs, _thread) =
+        let (edit_tool, project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "hello world\n"})).await;
         let (mut sender, input) = ToolInput::<EditFileToolInput>::test();
         let (event_stream, _receiver) = ToolCallEventStream::test();
@@ -1100,7 +969,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_input_partials_then_final(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "line 1\nline 2\nline 3\n"})).await;
         let (mut sender, input): (ToolInputSender, ToolInput<EditFileToolInput>) =
             ToolInput::test();
@@ -1140,7 +1009,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_input_sender_dropped_before_final(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "hello world\n"})).await;
         let (mut sender, input): (ToolInputSender, ToolInput<EditFileToolInput>) =
             ToolInput::test();
@@ -1157,67 +1026,6 @@ mod tests {
         assert!(
             result.is_err(),
             "Tool should error when sender is dropped without sending final input"
-        );
-    }
-
-    #[gpui::test]
-    async fn test_streaming_input_recv_drains_partials(cx: &mut TestAppContext) {
-        let (_edit_tool, write_tool, _project, _action_log, _fs, _thread) =
-            setup_test(cx, json!({"dir": {}})).await;
-        // Create a channel and send multiple partials before a final, then use
-        // ToolInput::resolved-style immediate delivery to confirm recv() works
-        // when partials are already buffered.
-        let (mut sender, input): (ToolInputSender, ToolInput<WriteFileToolInput>) =
-            ToolInput::test();
-        let (event_stream, _event_rx) = ToolCallEventStream::test();
-        let task = cx.update(|cx| write_tool.clone().run(input, event_stream, cx));
-
-        // Buffer several partials before sending the final
-        sender.send_partial(json!({}));
-        sender.send_partial(json!({"path": "root/dir/new.txt"}));
-        sender.send_partial(json!({
-            "path": "root/dir/new.txt",
-            "mode": "write"
-        }));
-        sender.send_full(json!({
-            "path": "root/dir/new.txt",
-            "mode": "write",
-            "content": "streamed content"
-        }));
-
-        let result = task.await;
-        let EditFileToolOutput::Success { new_text, .. } = result.unwrap() else {
-            panic!("expected success");
-        };
-        assert_eq!(new_text, "streamed content");
-    }
-
-    #[gpui::test]
-    async fn test_streaming_resolve_path_for_creating_file(cx: &mut TestAppContext) {
-        let mode = EditSessionMode::Write;
-
-        let result = test_resolve_path(&mode, "root/new.txt", cx);
-        assert_resolved_path_eq(result.await, rel_path("new.txt"));
-
-        let result = test_resolve_path(&mode, "new.txt", cx);
-        assert_resolved_path_eq(result.await, rel_path("new.txt"));
-
-        let result = test_resolve_path(&mode, "dir/new.txt", cx);
-        assert_resolved_path_eq(result.await, rel_path("dir/new.txt"));
-
-        let result = test_resolve_path(&mode, "root/dir/subdir/existing.txt", cx);
-        assert_resolved_path_eq(result.await, rel_path("dir/subdir/existing.txt"));
-
-        let result = test_resolve_path(&mode, "root/dir/subdir", cx);
-        assert_eq!(
-            result.await.unwrap_err(),
-            "Can't write to file: path is a directory"
-        );
-
-        let result = test_resolve_path(&mode, "root/dir/nonexistent_dir/new.txt", cx);
-        assert_eq!(
-            result.await.unwrap_err(),
-            "Can't create file: parent directory doesn't exist"
         );
     }
 
@@ -1274,277 +1082,8 @@ mod tests {
     }
 
     #[gpui::test]
-    async fn test_streaming_format_on_save(cx: &mut TestAppContext) {
-        init_test(cx);
-
-        let fs = project::FakeFs::new(cx.executor());
-        fs.insert_tree("/root", json!({"src": {}})).await;
-        let (_edit_tool, write_tool, project, action_log, fs, thread) =
-            setup_test_with_fs(cx, fs, &[path!("/root").as_ref()]).await;
-
-        let rust_language = Arc::new(language::Language::new(
-            language::LanguageConfig {
-                name: "Rust".into(),
-                matcher: language::LanguageMatcher {
-                    path_suffixes: vec!["rs".to_string()],
-                    ..Default::default()
-                },
-                ..Default::default()
-            },
-            None,
-        ));
-
-        let language_registry = project.read_with(cx, |project, _| project.languages().clone());
-        language_registry.add(rust_language);
-
-        let mut fake_language_servers = language_registry.register_fake_lsp(
-            "Rust",
-            language::FakeLspAdapter {
-                capabilities: lsp::ServerCapabilities {
-                    document_formatting_provider: Some(lsp::OneOf::Left(true)),
-                    ..Default::default()
-                },
-                ..Default::default()
-            },
-        );
-
-        fs.save(
-            path!("/root/src/main.rs").as_ref(),
-            &"initial content".into(),
-            language::LineEnding::Unix,
-        )
-        .await
-        .unwrap();
-
-        // Open the buffer to trigger LSP initialization
-        let buffer = project
-            .update(cx, |project, cx| {
-                project.open_local_buffer(path!("/root/src/main.rs"), cx)
-            })
-            .await
-            .unwrap();
-
-        // Register the buffer with language servers
-        let _handle = project.update(cx, |project, cx| {
-            project.register_buffer_with_language_servers(&buffer, cx)
-        });
-
-        const UNFORMATTED_CONTENT: &str = "fn main() {println!(\"Hello!\");}\
-";
-        const FORMATTED_CONTENT: &str = "This file was formatted by the fake formatter in the test.\
-";
-
-        // Get the fake language server and set up formatting handler
-        let fake_language_server = fake_language_servers.next().await.unwrap();
-        fake_language_server.set_request_handler::<lsp::request::Formatting, _, _>({
-            |_, _| async move {
-                Ok(Some(vec![lsp::TextEdit {
-                    range: lsp::Range::new(lsp::Position::new(0, 0), lsp::Position::new(1, 0)),
-                    new_text: FORMATTED_CONTENT.to_string(),
-                }]))
-            }
-        });
-
-        // Test with format_on_save enabled
-        cx.update(|cx| {
-            SettingsStore::update_global(cx, |store, cx| {
-                store.update_user_settings(cx, |settings| {
-                    settings.project.all_languages.defaults.format_on_save = Some(FormatOnSave::On);
-                    settings.project.all_languages.defaults.formatter =
-                        Some(language::language_settings::FormatterList::default());
-                });
-            });
-        });
-
-        // Use streaming pattern so executor can pump the LSP request/response
-        let (mut sender, input) = ToolInput::<WriteFileToolInput>::test();
-        let (event_stream, _receiver) = ToolCallEventStream::test();
-
-        let task = cx.update(|cx| write_tool.clone().run(input, event_stream, cx));
-
-        sender.send_partial(json!({
-            "path": "root/src/main.rs",
-        }));
-        cx.run_until_parked();
-
-        sender.send_full(json!({
-            "path": "root/src/main.rs",
-            "content": UNFORMATTED_CONTENT
-        }));
-
-        let result = task.await;
-        assert!(result.is_ok());
-
-        cx.executor().run_until_parked();
-
-        let new_content = fs.load(path!("/root/src/main.rs").as_ref()).await.unwrap();
-        assert_eq!(
-            new_content.replace("\r\n", "\n"),
-            FORMATTED_CONTENT,
-            "Code should be formatted when format_on_save is enabled"
-        );
-
-        let stale_buffer_count = thread
-            .read_with(cx, |thread, _cx| thread.action_log.clone())
-            .read_with(cx, |log, cx| log.stale_buffers(cx).count());
-
-        assert_eq!(
-            stale_buffer_count, 0,
-            "BUG: Buffer is incorrectly marked as stale after format-on-save. Found {} stale buffers.",
-            stale_buffer_count
-        );
-
-        // Test with format_on_save disabled
-        cx.update(|cx| {
-            SettingsStore::update_global(cx, |store, cx| {
-                store.update_user_settings(cx, |settings| {
-                    settings.project.all_languages.defaults.format_on_save =
-                        Some(FormatOnSave::Off);
-                });
-            });
-        });
-
-        let (mut sender, input) = ToolInput::<WriteFileToolInput>::test();
-        let (event_stream, _receiver) = ToolCallEventStream::test();
-
-        let tool2 = Arc::new(WriteFileTool::new(
-            project.clone(),
-            thread.downgrade(),
-            action_log.clone(),
-            language_registry,
-        ));
-
-        let task = cx.update(|cx| tool2.run(input, event_stream, cx));
-
-        sender.send_partial(json!({
-            "path": "root/src/main.rs",
-        }));
-        cx.run_until_parked();
-
-        sender.send_full(json!({
-            "path": "root/src/main.rs",
-            "content": UNFORMATTED_CONTENT
-        }));
-
-        let result = task.await;
-        assert!(result.is_ok());
-
-        cx.executor().run_until_parked();
-
-        let new_content = fs.load(path!("/root/src/main.rs").as_ref()).await.unwrap();
-        assert_eq!(
-            new_content.replace("\r\n", "\n"),
-            UNFORMATTED_CONTENT,
-            "Code should not be formatted when format_on_save is disabled"
-        );
-    }
-
-    #[gpui::test]
-    async fn test_streaming_remove_trailing_whitespace(cx: &mut TestAppContext) {
-        init_test(cx);
-
-        let fs = project::FakeFs::new(cx.executor());
-        fs.insert_tree("/root", json!({"src": {}})).await;
-        fs.save(
-            path!("/root/src/main.rs").as_ref(),
-            &"initial content".into(),
-            language::LineEnding::Unix,
-        )
-        .await
-        .unwrap();
-        let (_edit_tool, write_tool, project, action_log, fs, thread) =
-            setup_test_with_fs(cx, fs, &[path!("/root").as_ref()]).await;
-        let language_registry = project.read_with(cx, |p, _cx| p.languages().clone());
-
-        // Test with remove_trailing_whitespace_on_save enabled
-        cx.update(|cx| {
-            SettingsStore::update_global(cx, |store, cx| {
-                store.update_user_settings(cx, |settings| {
-                    settings
-                        .project
-                        .all_languages
-                        .defaults
-                        .remove_trailing_whitespace_on_save = Some(true);
-                });
-            });
-        });
-
-        const CONTENT_WITH_TRAILING_WHITESPACE: &str =
-            "fn main() {  \n    println!(\"Hello!\");  \n}\n";
-
-        let result = cx
-            .update(|cx| {
-                write_tool.clone().run(
-                    ToolInput::resolved(WriteFileToolInput {
-                        path: "root/src/main.rs".into(),
-                        content: CONTENT_WITH_TRAILING_WHITESPACE.into(),
-                    }),
-                    ToolCallEventStream::test().0,
-                    cx,
-                )
-            })
-            .await;
-        assert!(result.is_ok());
-
-        cx.executor().run_until_parked();
-
-        assert_eq!(
-            fs.load(path!("/root/src/main.rs").as_ref())
-                .await
-                .unwrap()
-                .replace("\r\n", "\n"),
-            "fn main() {\n    println!(\"Hello!\");\n}\n",
-            "Trailing whitespace should be removed when remove_trailing_whitespace_on_save is enabled"
-        );
-
-        // Test with remove_trailing_whitespace_on_save disabled
-        cx.update(|cx| {
-            SettingsStore::update_global(cx, |store, cx| {
-                store.update_user_settings(cx, |settings| {
-                    settings
-                        .project
-                        .all_languages
-                        .defaults
-                        .remove_trailing_whitespace_on_save = Some(false);
-                });
-            });
-        });
-
-        let tool2 = Arc::new(WriteFileTool::new(
-            project.clone(),
-            thread.downgrade(),
-            action_log.clone(),
-            language_registry,
-        ));
-
-        let result = cx
-            .update(|cx| {
-                tool2.run(
-                    ToolInput::resolved(WriteFileToolInput {
-                        path: "root/src/main.rs".into(),
-                        content: CONTENT_WITH_TRAILING_WHITESPACE.into(),
-                    }),
-                    ToolCallEventStream::test().0,
-                    cx,
-                )
-            })
-            .await;
-        assert!(result.is_ok());
-
-        cx.executor().run_until_parked();
-
-        let final_content = fs.load(path!("/root/src/main.rs").as_ref()).await.unwrap();
-        assert_eq!(
-            final_content.replace("\r\n", "\n"),
-            CONTENT_WITH_TRAILING_WHITESPACE,
-            "Trailing whitespace should remain when remove_trailing_whitespace_on_save is disabled"
-        );
-    }
-
-    #[gpui::test]
     async fn test_streaming_authorize(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
-            setup_test(cx, json!({})).await;
+        let (edit_tool, _project, _action_log, _fs, _thread) = setup_test(cx, json!({})).await;
 
         // Test 1: Path with .zed component should require confirmation
         let (stream_tx, mut stream_rx) = ToolCallEventStream::test();
@@ -1645,7 +1184,7 @@ mod tests {
         fs.insert_tree("/outside", json!({})).await;
         fs.insert_symlink("/root/link", PathBuf::from("/outside"))
             .await;
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test_with_fs(cx, fs, &[path!("/root").as_ref()]).await;
 
         cx.update(|cx| {
@@ -1706,7 +1245,7 @@ mod tests {
         )
         .await
         .unwrap();
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test_with_fs(cx, fs, &[path!("/root").as_ref()]).await;
 
         let (stream_tx, mut stream_rx) = ToolCallEventStream::test();
@@ -1751,7 +1290,7 @@ mod tests {
         )
         .await
         .unwrap();
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test_with_fs(cx, fs, &[path!("/root").as_ref()]).await;
 
         let (stream_tx, mut stream_rx) = ToolCallEventStream::test();
@@ -1806,7 +1345,7 @@ mod tests {
         )
         .await
         .unwrap();
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test_with_fs(cx, fs, &[path!("/root").as_ref()]).await;
 
         let (stream_tx, mut stream_rx) = ToolCallEventStream::test();
@@ -1835,7 +1374,7 @@ mod tests {
         init_test(cx);
         let fs = project::FakeFs::new(cx.executor());
         fs.insert_tree("/project", json!({})).await;
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test_with_fs(cx, fs, &[path!("/project").as_ref()]).await;
 
         let test_cases = vec![
@@ -1905,7 +1444,7 @@ mod tests {
             }),
         )
         .await;
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) = setup_test_with_fs(
+        let (edit_tool, _project, _action_log, _fs, _thread) = setup_test_with_fs(
             cx,
             fs,
             &[
@@ -1968,7 +1507,7 @@ mod tests {
             }),
         )
         .await;
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test_with_fs(cx, fs, &[path!("/project").as_ref()]).await;
 
         let test_cases = vec![
@@ -2024,7 +1563,7 @@ mod tests {
             }),
         )
         .await;
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test_with_fs(cx, fs, &[path!("/project").as_ref()]).await;
 
         let modes = vec![EditSessionMode::Edit, EditSessionMode::Write];
@@ -2062,7 +1601,7 @@ mod tests {
         init_test(cx);
         let fs = project::FakeFs::new(cx.executor());
         fs.insert_tree("/project", json!({})).await;
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test_with_fs(cx, fs, &[path!("/project").as_ref()]).await;
 
         cx.update(|cx| {
@@ -2092,66 +1631,8 @@ mod tests {
     }
 
     #[gpui::test]
-    async fn test_streaming_diff_finalization(cx: &mut TestAppContext) {
-        init_test(cx);
-        let fs = project::FakeFs::new(cx.executor());
-        fs.insert_tree("/", json!({"main.rs": ""})).await;
-        let (_edit_tool, write_tool, project, action_log, _fs, thread) =
-            setup_test_with_fs(cx, fs, &[path!("/").as_ref()]).await;
-        let language_registry = project.read_with(cx, |p, _cx| p.languages().clone());
-
-        // Ensure the diff is finalized after the edit completes.
-        {
-            let (stream_tx, mut stream_rx) = ToolCallEventStream::test();
-            let edit = cx.update(|cx| {
-                write_tool.clone().run(
-                    ToolInput::resolved(WriteFileToolInput {
-                        path: path!("/main.rs").into(),
-                        content: "new content".into(),
-                    }),
-                    stream_tx,
-                    cx,
-                )
-            });
-            stream_rx.expect_update_fields().await;
-            let diff = stream_rx.expect_diff().await;
-            diff.read_with(cx, |diff, _| assert!(matches!(diff, Diff::Pending(_))));
-            cx.run_until_parked();
-            edit.await.unwrap();
-            diff.read_with(cx, |diff, _| assert!(matches!(diff, Diff::Finalized(_))));
-        }
-
-        // Ensure the diff is finalized if the tool call gets dropped.
-        {
-            let tool = Arc::new(WriteFileTool::new(
-                project.clone(),
-                thread.downgrade(),
-                action_log,
-                language_registry,
-            ));
-            let (stream_tx, mut stream_rx) = ToolCallEventStream::test();
-            let edit = cx.update(|cx| {
-                tool.run(
-                    ToolInput::resolved(WriteFileToolInput {
-                        path: path!("/main.rs").into(),
-                        content: "dropped content".into(),
-                    }),
-                    stream_tx,
-                    cx,
-                )
-            });
-            stream_rx.expect_update_fields().await;
-            let diff = stream_rx.expect_diff().await;
-            diff.read_with(cx, |diff, _| assert!(matches!(diff, Diff::Pending(_))));
-            drop(edit);
-            cx.run_until_parked();
-            diff.read_with(cx, |diff, _| assert!(matches!(diff, Diff::Finalized(_))));
-        }
-    }
-
-    #[gpui::test]
     async fn test_streaming_consecutive_edits_work(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, project, action_log, _fs, _thread) =
+        let (edit_tool, project, action_log, _fs, _thread) =
             setup_test(cx, json!({"test.txt": "original content"})).await;
         let read_tool = Arc::new(crate::ReadFileTool::new(
             project.clone(),
@@ -2221,7 +1702,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_external_modification_matching_edit_succeeds(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, project, action_log, fs, _thread) =
+        let (edit_tool, project, action_log, fs, _thread) =
             setup_test(cx, json!({"test.txt": "original content"})).await;
         let read_tool = Arc::new(crate::ReadFileTool::new(
             project.clone(),
@@ -2306,7 +1787,7 @@ mod tests {
     async fn test_streaming_external_modification_mentioned_when_match_fails(
         cx: &mut TestAppContext,
     ) {
-        let (edit_tool, _write_tool, project, action_log, fs, _thread) =
+        let (edit_tool, project, action_log, fs, _thread) =
             setup_test(cx, json!({"test.txt": "original content"})).await;
         let read_tool = Arc::new(crate::ReadFileTool::new(
             project.clone(),
@@ -2393,7 +1874,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_dirty_buffer_detected(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, project, action_log, _fs, _thread) =
+        let (edit_tool, project, action_log, _fs, _thread) =
             setup_test(cx, json!({"test.txt": "original content"})).await;
         let read_tool = Arc::new(crate::ReadFileTool::new(
             project.clone(),
@@ -2485,7 +1966,7 @@ mod tests {
         // old_text as a substring. Because edits resolve sequentially
         // against the current buffer, edit 2 finds a unique match in
         // the modified buffer and succeeds.
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "aaa\nbbb\nccc\nddd\neee\n"})).await;
         let (mut sender, input) = ToolInput::<EditFileToolInput>::test();
         let (event_stream, _receiver) = ToolCallEventStream::test();
@@ -2533,210 +2014,8 @@ mod tests {
     }
 
     #[gpui::test]
-    async fn test_streaming_create_content_streamed(cx: &mut TestAppContext) {
-        let (_edit_tool, write_tool, project, _action_log, _fs, _thread) =
-            setup_test(cx, json!({"dir": {}})).await;
-        let (mut sender, input) = ToolInput::<WriteFileToolInput>::test();
-        let (event_stream, _receiver) = ToolCallEventStream::test();
-        let task = cx.update(|cx| write_tool.clone().run(input, event_stream, cx));
-
-        // Transition to BufferResolved
-        sender.send_partial(json!({
-            "path": "root/dir/new_file.txt",
-            "mode": "write"
-        }));
-        cx.run_until_parked();
-
-        // Stream content incrementally
-        sender.send_partial(json!({
-            "path": "root/dir/new_file.txt",
-            "mode": "write",
-            "content": "line 1\n"
-        }));
-        cx.run_until_parked();
-
-        // Verify buffer has partial content
-        let buffer = project.update(cx, |project, cx| {
-            let path = project
-                .find_project_path("root/dir/new_file.txt", cx)
-                .unwrap();
-            project.get_open_buffer(&path, cx).unwrap()
-        });
-        assert_eq!(buffer.read_with(cx, |b, _| b.text()), "line 1\n");
-
-        // Stream more content
-        sender.send_partial(json!({
-            "path": "root/dir/new_file.txt",
-            "mode": "write",
-            "content": "line 1\nline 2\n"
-        }));
-        cx.run_until_parked();
-        assert_eq!(buffer.read_with(cx, |b, _| b.text()), "line 1\nline 2\n");
-
-        // Stream final chunk
-        sender.send_partial(json!({
-            "path": "root/dir/new_file.txt",
-            "mode": "write",
-            "content": "line 1\nline 2\nline 3\n"
-        }));
-        cx.run_until_parked();
-        assert_eq!(
-            buffer.read_with(cx, |b, _| b.text()),
-            "line 1\nline 2\nline 3\n"
-        );
-
-        // Send final input
-        sender.send_full(json!({
-            "path": "root/dir/new_file.txt",
-            "mode": "write",
-            "content": "line 1\nline 2\nline 3\n"
-        }));
-
-        let result = task.await;
-        let EditFileToolOutput::Success { new_text, .. } = result.unwrap() else {
-            panic!("expected success");
-        };
-        assert_eq!(new_text, "line 1\nline 2\nline 3\n");
-    }
-
-    #[gpui::test]
-    async fn test_streaming_overwrite_diff_revealed_during_streaming(cx: &mut TestAppContext) {
-        let (_edit_tool, write_tool, _project, _action_log, _fs, _thread) = setup_test(
-            cx,
-            json!({"file.txt": "old line 1\nold line 2\nold line 3\n"}),
-        )
-        .await;
-        let (mut sender, input) = ToolInput::<WriteFileToolInput>::test();
-        let (event_stream, mut receiver) = ToolCallEventStream::test();
-        let task = cx.update(|cx| write_tool.clone().run(input, event_stream, cx));
-
-        // Transition to BufferResolved
-        sender.send_partial(json!({
-            "path": "root/file.txt",
-        }));
-        cx.run_until_parked();
-
-        sender.send_partial(json!({
-            "path": "root/file.txt",
-            "mode": "write"
-        }));
-        cx.run_until_parked();
-
-        // Get the diff entity from the event stream
-        receiver.expect_update_fields().await;
-        let diff = receiver.expect_diff().await;
-
-        // Diff starts pending with no revealed ranges
-        diff.read_with(cx, |diff, cx| {
-            assert!(matches!(diff, Diff::Pending(_)));
-            assert!(!diff.has_revealed_range(cx));
-        });
-
-        // Stream first content chunk
-        sender.send_partial(json!({
-            "path": "root/file.txt",
-            "mode": "write",
-            "content": "new line 1\n"
-        }));
-        cx.run_until_parked();
-
-        // Diff should now have revealed ranges showing the new content
-        diff.read_with(cx, |diff, cx| {
-            assert!(diff.has_revealed_range(cx));
-        });
-
-        // Send final input
-        sender.send_full(json!({
-            "path": "root/file.txt",
-            "mode": "write",
-            "content": "new line 1\nnew line 2\n"
-        }));
-
-        let result = task.await;
-        let EditFileToolOutput::Success {
-            new_text, old_text, ..
-        } = result.unwrap()
-        else {
-            panic!("expected success");
-        };
-        assert_eq!(new_text, "new line 1\nnew line 2\n");
-        assert_eq!(*old_text, "old line 1\nold line 2\nold line 3\n");
-
-        // Diff is finalized after completion
-        diff.read_with(cx, |diff, _| assert!(matches!(diff, Diff::Finalized(_))));
-    }
-
-    #[gpui::test]
-    async fn test_streaming_overwrite_content_streamed(cx: &mut TestAppContext) {
-        let (_edit_tool, write_tool, project, _action_log, _fs, _thread) = setup_test(
-            cx,
-            json!({"file.txt": "old line 1\nold line 2\nold line 3\n"}),
-        )
-        .await;
-        let (mut sender, input) = ToolInput::<WriteFileToolInput>::test();
-        let (event_stream, _receiver) = ToolCallEventStream::test();
-        let task = cx.update(|cx| write_tool.clone().run(input, event_stream, cx));
-
-        // Transition to BufferResolved
-        sender.send_partial(json!({
-            "path": "root/file.txt",
-            "mode": "write"
-        }));
-        cx.run_until_parked();
-
-        // Verify buffer still has old content (no content partial yet)
-        let buffer = project.update(cx, |project, cx| {
-            let path = project.find_project_path("root/file.txt", cx).unwrap();
-            project.open_buffer(path, cx)
-        });
-        let buffer = buffer.await.unwrap();
-        assert_eq!(
-            buffer.read_with(cx, |b, _| b.text()),
-            "old line 1\nold line 2\nold line 3\n"
-        );
-
-        // First content partial replaces old content
-        sender.send_partial(json!({
-            "path": "root/file.txt",
-            "mode": "write",
-            "content": "new line 1\n"
-        }));
-        cx.run_until_parked();
-        assert_eq!(buffer.read_with(cx, |b, _| b.text()), "new line 1\n");
-
-        // Subsequent content partials append
-        sender.send_partial(json!({
-            "path": "root/file.txt",
-            "mode": "write",
-            "content": "new line 1\nnew line 2\n"
-        }));
-        cx.run_until_parked();
-        assert_eq!(
-            buffer.read_with(cx, |b, _| b.text()),
-            "new line 1\nnew line 2\n"
-        );
-
-        // Send final input with complete content
-        sender.send_full(json!({
-            "path": "root/file.txt",
-            "mode": "write",
-            "content": "new line 1\nnew line 2\nnew line 3\n"
-        }));
-
-        let result = task.await;
-        let EditFileToolOutput::Success {
-            new_text, old_text, ..
-        } = result.unwrap()
-        else {
-            panic!("expected success");
-        };
-        assert_eq!(new_text, "new line 1\nnew line 2\nnew line 3\n");
-        assert_eq!(*old_text, "old line 1\nold line 2\nold line 3\n");
-    }
-
-    #[gpui::test]
     async fn test_streaming_edit_json_fixer_escape_corruption(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "hello\nworld\nfoo\n"})).await;
         let (mut sender, input) = ToolInput::<EditFileToolInput>::test();
         let (event_stream, _receiver) = ToolCallEventStream::test();
@@ -2784,7 +2063,7 @@ mod tests {
 
     #[gpui::test]
     async fn test_streaming_final_input_stringified_edits_succeeds(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "hello\nworld\n"})).await;
         let (mut sender, input) = ToolInput::<EditFileToolInput>::test();
         let (event_stream, _receiver) = ToolCallEventStream::test();
@@ -2813,7 +2092,7 @@ mod tests {
     // reports changed buffers so that the Accept All / Reject All review UI appears.
     #[gpui::test]
     async fn test_streaming_edit_file_tool_registers_changed_buffers(cx: &mut TestAppContext) {
-        let (edit_tool, _write_tool, _project, action_log, _fs, _thread) =
+        let (edit_tool, _project, action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "line 1\nline 2\nline 3\n"})).await;
         cx.update(|cx| {
             let mut settings = agent_settings::AgentSettings::get_global(cx).clone();
@@ -2850,90 +2129,12 @@ mod tests {
     }
 
     // Same test but for Write mode (overwrite entire file).
-    #[gpui::test]
-    async fn test_streaming_edit_file_tool_write_mode_registers_changed_buffers(
-        cx: &mut TestAppContext,
-    ) {
-        let (_edit_tool, write_tool, _project, action_log, _fs, _thread) =
-            setup_test(cx, json!({"file.txt": "original content"})).await;
-        cx.update(|cx| {
-            let mut settings = agent_settings::AgentSettings::get_global(cx).clone();
-            settings.tool_permissions.default = settings::ToolPermissionMode::Allow;
-            agent_settings::AgentSettings::override_global(settings, cx);
-        });
-
-        let (event_stream, _rx) = ToolCallEventStream::test();
-        let task = cx.update(|cx| {
-            write_tool.clone().run(
-                ToolInput::resolved(WriteFileToolInput {
-                    path: "root/file.txt".into(),
-                    content: "completely new content".into(),
-                }),
-                event_stream,
-                cx,
-            )
-        });
-
-        let result = task.await;
-        assert!(result.is_ok(), "write should succeed: {:?}", result.err());
-
-        cx.run_until_parked();
-
-        let changed = action_log.read_with(cx, |log, cx| log.changed_buffers(cx));
-        assert!(
-            !changed.is_empty(),
-            "action_log.changed_buffers() should be non-empty after streaming write, \
-             but no changed buffers were found \u{2014} Accept All / Reject All will not appear"
-        );
-    }
-
-    #[gpui::test]
-    async fn test_streaming_edit_file_tool_fields_out_of_order_in_write_mode(
-        cx: &mut TestAppContext,
-    ) {
-        let (_edit_tool, write_tool, _project, _action_log, _fs, _thread) =
-            setup_test(cx, json!({"file.txt": "old_content"})).await;
-        let (mut sender, input) = ToolInput::<WriteFileToolInput>::test();
-        let (event_stream, _receiver) = ToolCallEventStream::test();
-        let task = cx.update(|cx| write_tool.clone().run(input, event_stream, cx));
-
-        sender.send_partial(json!({
-            "mode": "write"
-        }));
-        cx.run_until_parked();
-
-        sender.send_partial(json!({
-            "mode": "write",
-            "content": "new_content"
-        }));
-        cx.run_until_parked();
-
-        sender.send_partial(json!({
-            "mode": "write",
-            "content": "new_content",
-            "path": "root"
-        }));
-        cx.run_until_parked();
-
-        // Send final.
-        sender.send_full(json!({
-            "mode": "write",
-            "content": "new_content",
-            "path": "root/file.txt"
-        }));
-
-        let result = task.await;
-        let EditFileToolOutput::Success { new_text, .. } = result.unwrap() else {
-            panic!("expected success");
-        };
-        assert_eq!(new_text, "new_content");
-    }
 
     #[gpui::test]
     async fn test_streaming_edit_file_tool_fields_out_of_order_in_edit_mode(
         cx: &mut TestAppContext,
     ) {
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.txt": "old_content"})).await;
         let (mut sender, input) = ToolInput::<EditFileToolInput>::test();
         let (event_stream, _receiver) = ToolCallEventStream::test();
@@ -2993,7 +2194,7 @@ mod tests {
         "#}
         .to_string();
 
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.rs": file_content})).await;
 
         // The model sends old_text with a PARTIAL last line.
@@ -3038,7 +2239,7 @@ mod tests {
         let new_text = "one\ntwo\ntarget\n";
         let expected = "before\none\ntwo\ntarget\n\nafter\n";
 
-        let (edit_tool, _write_tool, _project, _action_log, _fs, _thread) =
+        let (edit_tool, _project, _action_log, _fs, _thread) =
             setup_test(cx, json!({"file.rs": file_content})).await;
         let (mut sender, input) = ToolInput::<EditFileToolInput>::test();
         let (event_stream, _receiver) = ToolCallEventStream::test();
@@ -3064,55 +2265,6 @@ mod tests {
             final_text,
             expected,
             "Edit should preserve a single blank line before test_after"
-        );
-    }
-
-    #[gpui::test]
-    async fn test_streaming_reject_created_file_deletes_it(cx: &mut TestAppContext) {
-        let (_edit_tool, write_tool, _project, action_log, fs, _thread) =
-            setup_test(cx, json!({"dir": {}})).await;
-        cx.update(|cx| {
-            let mut settings = agent_settings::AgentSettings::get_global(cx).clone();
-            settings.tool_permissions.default = settings::ToolPermissionMode::Allow;
-            agent_settings::AgentSettings::override_global(settings, cx);
-        });
-
-        // Create a new file via the streaming edit file tool
-        let (event_stream, _rx) = ToolCallEventStream::test();
-        let task = cx.update(|cx| {
-            write_tool.clone().run(
-                ToolInput::resolved(WriteFileToolInput {
-                    path: "root/dir/new_file.txt".into(),
-                    content: "Hello, World!".into(),
-                }),
-                event_stream,
-                cx,
-            )
-        });
-        let result = task.await;
-        assert!(result.is_ok(), "create should succeed: {:?}", result.err());
-        cx.run_until_parked();
-
-        assert!(
-            fs.is_file(path!("/root/dir/new_file.txt").as_ref()).await,
-            "file should exist after creation"
-        );
-
-        // Reject all edits — this should delete the newly created file
-        let changed = action_log.read_with(cx, |log, cx| log.changed_buffers(cx));
-        assert!(
-            !changed.is_empty(),
-            "action_log should track the created file as changed"
-        );
-
-        action_log
-            .update(cx, |log, cx| log.reject_all_edits(None, cx))
-            .await;
-        cx.run_until_parked();
-
-        assert!(
-            !fs.is_file(path!("/root/dir/new_file.txt").as_ref()).await,
-            "file should be deleted after rejecting creation, but an empty file was left behind"
         );
     }
 
@@ -3159,7 +2311,6 @@ mod tests {
         worktree_paths: &[&std::path::Path],
     ) -> (
         Arc<EditFileTool>,
-        Arc<WriteFileTool>,
         Entity<Project>,
         Entity<ActionLog>,
         Arc<project::FakeFs>,
@@ -3187,13 +2338,7 @@ mod tests {
             action_log.clone(),
             language_registry.clone(),
         ));
-        let write_tool = Arc::new(WriteFileTool::new(
-            project.clone(),
-            thread.downgrade(),
-            action_log.clone(),
-            language_registry,
-        ));
-        (edit_tool, write_tool, project, action_log, fs, thread)
+        (edit_tool, project, action_log, fs, thread)
     }
 
     async fn setup_test(
@@ -3201,7 +2346,6 @@ mod tests {
         initial_tree: serde_json::Value,
     ) -> (
         Arc<EditFileTool>,
-        Arc<WriteFileTool>,
         Entity<Project>,
         Entity<ActionLog>,
         Arc<project::FakeFs>,
